requirements.txt
# DeciBuy v2.0 — Requirements
# Install with: pip install -r requirements.txt

# ── Core UI ──
streamlit>=1.32.0

# ── Data processing ──
pandas>=2.0.0
numpy>=1.24.0

# ── Visualization ──
plotly>=5.18.0

# ── Machine Learning ──
scikit-learn>=1.3.0

# ── OCR (install at least one) ──
# Option A — EasyOCR (recommended, no system install needed):
easyocr>=1.7.0
# Option B — pytesseract (requires Tesseract installed on OS):
# pytesseract>=0.3.10
Pillow>=10.0.0

# ── AI (Ollama runs separately, no pip package needed) ──
# Install Ollama from: https://ollama.com
# Then run: ollama pull qwen3-vl:4b
# (or any other model you prefer)
